Code generator bases on bound results to choose fixed-sized integer type for all variables (including parameters) in a function

+++ tests/bound/ifelse/naive_bound/naive/ifelse.c
@@ -1,11 +1,11 @@
 #include "ifelse.h"
-int64_t _f_(int64_t x){
+int64_t _f_(uint16_t x){
 	int64_t r = 0;
 	uint16_t _2 = 0;
-	int64_t _3 = 0;
+	uint16_t _3 = 0;
 	uint16_t _4 = 0;
 	uint16_t _5 = 0;
-	int64_t _6 = 0;
+	uint16_t _6 = 0;
 	//const %2 = 10 : int
 	_2 = 10;
 	//ifge %0, %2 goto blklab1 : int
