+++ tests/bound/ifelse/naive_bound/naive/ifelse.h
@@ -1,4 +1,4 @@
 #include "Util.h"
 #include "WyRT.h"
-int64_t _f_(int64_t x);
+int64_t _f_(uint16_t x);
 int main(int argc, char** args);
