+++ tests/bound/whileloop/naive_bound/naive/whileloop.h
@@ -1,4 +1,4 @@
 #include "Util.h"
 #include "WyRT.h"
-int64_t _f_(int64_t limit);
+int64_t _f_(uint32_t limit);
 int main(int argc, char** args);
